requirements.txt
click==8.0.1
Faker==8.8.1
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==3.0.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.26.1
pandas==2.1.1
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==1.4.18
text-unidecode==1.3
typing_extensions==4.8.0
tzdata==2023.3
Werkzeug==2.0.1
